requirements.txt
streamlit==1.41.1
sentence-transformers==3.3.1
faiss-cpu==1.7.4
beautifulsoup4==4.12.3
requests==2.32.3
numpy==1.24.3
scikit-learn==1.6.0
scipy==1.14.1
torch==2.5.1
tqdm==4.67.1
PyYAML==6.0.2
transformers
pandas
